attenuation issue fixed
it was a problem in the SetFloat function in shader manager because i was casting the value to an int!!!! i really outdo myself sometimes

+++ GoofyGL/shader_manager.cpp
@@ -97,7 +97,7 @@ void Shader::SetInt(const std::string& name, int value) const
 
 void Shader::SetFloat(const std::string& name, float value) const
 {
-	glUniform1f(glGetUniformLocation(ID, name.c_str()), (int)value);
+	glUniform1f(glGetUniformLocation(ID, name.c_str()), value);
 }
 
 void Shader::SetMat4(const std::string& name, const glm::mat4& mat) const
